+++ desktop/cmp/filechooser/FileChooserModel.ts
@@ -9,6 +9,7 @@ import {HoistModel, managed} from '@xh/hoist/core';
 import {actionCol, calcActionColWidth} from '@xh/hoist/desktop/cmp/grid';
 import '@xh/hoist/desktop/register';
 import {Icon} from '@xh/hoist/icon';
+import {FileRejection} from '@xh/hoist/kit/react-dropzone';
 import {action, makeObservable, observable} from '@xh/hoist/mobx';
 import {filesize} from 'filesize';
 import {find, isEmpty, uniqBy, without} from 'lodash';
@@ -118,7 +119,7 @@ export class FileChooserModel extends HoistModel {
     }
 
     @action
-    onDrop(accepted, rejected, enableMulti) {
+    onDrop(accepted: File[], rejected: readonly FileRejection[], enableMulti: boolean) {
         if (!isEmpty(accepted)) {
             if (!enableMulti) {
                 this.setSingleFile(accepted[0]);